requirements.txt
albumentations
cloths-segmentation
iglovikov-helper-functions
matplotlib
numpy
opencv-python-headless
Pillow
pycocotools
torch
python-telegram-bot